refactor(layout): MobileHeader sous 300 lignes — éclatement complet (H12)
- useMobileHeader.js(emit): détection mobile, panneaux exclusifs, infos user,
  chemin profil par rôle, relais toggle-sidebar, déconnexion
- MobileNotificationsPanel/MobileUserMenuPanel: panneaux présentationnels
  (visibles selon show, émettent close/logout); chrome CSS dupliqué VERBATIM
  dans chaque panneau (choix: pas de partial), avatars fusionnés base+spécifique
- MobileHeader.vue: orchestrateur 201 lignes (433 → 201)
- Tests: useMobileHeader (logique) + montage des 2 panneaux; MobileHeader.test conservé
